% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/circleFn.R
\name{circleVertices}
\alias{circleVertices}
\title{Generate vertex coordinates for a circle.}
\usage{
circleVertices(xc, yc, r, npoints = 25)
}
\arguments{
\item{xc}{centre X}

\item{yc}{centre Y}

\item{r}{radius}

\item{npoints}{number of vertices}
}
\value{
a 2-column matrix of X and Y values
}
\description{
Generates vertex coordinates for a circle with given centre and radius.
}
\examples{
\dontrun{

## draw some random circles with ggplot
library(ggplot2)

xmax = 100
ymax = 200
rmax <- 20
N <- 100

## Random centre coordinates and sizes
xyr <- data.frame(x=runif(N, 0, xbound),
                  y=runif(N, 0, ybound),
                  r=rbeta(N, 1, 5) * rmax)

## Generate vertices for each circle and accumulate in
## a single data.frame with an id col
dat <- do.call("rbind", lapply(1:nrow(m), function(i) {
  df <- as.data.frame(circleVertices(m[i,1], m[i,2], m[i,3], 25))
  df$id <- i
  df
}))

## Draw them
ggplot(dat, aes(x, y, group=id)) +
  geom_path(colour="steelblue") +
  coord_equal(xlim=c(0, xbound), ylim=c(0, ybound)) +
  theme_bw()
}
}

