% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expasym.R
\name{expasym}
\alias{expasym}
\title{Area transformation function}
\usage{
expasym(x, a, b)
}
\arguments{
\item{x}{input area value(s)}

\item{a}{coefficient}

\item{b}{coefficient}
}
\description{
TODO - where do we use this ?
}
\details{
Function has the form:
\preformatted{area' = a * (1 - exp(-b * area))}
}

