% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpeciesParams-class.R
\docType{methods}
\name{getFunctionOrElse,SpeciesParams-method}
\alias{getFunctionOrElse,SpeciesParams-method}
\title{Gets an optional function if present, or a fallback function object.}
\usage{
\S4method{getFunctionOrElse}{SpeciesParams}(object, fnName, fallback)
}
\arguments{
\item{object}{\code{SpeciesParam} object}

\item{fnName}{name of the function slot (partial matching is used)}

\item{fallback}{fallback function object}
}
\value{
the optional function if present, or the fallback
}
\description{
Gets an optional function if present, or a fallback function object.
}

