% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{iterate_layout}
\alias{iterate_layout}
\title{Attempts to position circles without overlap.}
\usage{
iterate_layout(xyr, xbound, ybound, maxiter)
}
\arguments{
\item{xyr}{3 column matrix (centre x, centre y, radius)}

\item{xbound}{limit in X direction}

\item{ybound}{limit in Y direction}

\item{maxiter}{maximum number of iterations}
}
\value{
the number of iterations performed.
}
\description{
Given an input matrix of circle positions and sizes, attempts to position them
without overlap by iterating the pair-repulsion algorithm.
}

