% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/packCircles.R
\name{packCircles}
\alias{packCircles}
\title{Find non-overlapping configuration of circles.}
\usage{
packCircles(xyr, xbound, ybound, max.iter)
}
\arguments{
\item{xyr}{3-column matrix or data.frame (centre X, centre Y, radius)}

\item{xbound}{X upper limit (lower assumed 0)}

\item{ybound}{Y upper limit (lower assumed 0)}

\item{max.iter}{maximum number of iterations to attempt}
}
\value{
A list with components:
  \describe{
    \item{config}{A 3-column matrix or data.frame (centre x, centre y, radius).}
    \item{n.iter}{Number of iterations performed.}
  }
}
\description{
Attempts to find a configuration for a given set of circles
in a rectangle where there is no overlap between circles.
}

