% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpeciesParams-class.R
\docType{class}
\name{SpeciesParams-class}
\alias{SpeciesParams-class}
\title{Class to hold species parameters used in simulation.}
\description{
\code{SpeciesParams} is an S4 class which defines the species parameters used
by the \code{tm.site} simulation. This replaces the previous use of a named list
for the same purpose and provides some type and validity checking of arguments.
Note that at the moment there are no associated methods to get and set elements
(as is usually the case for S4 classes), instead you deal with the slots directly.
Given that this is just a data class that shouldn't be too painful.
}
\section{Slots}{

\describe{
\item{\code{name}}{Species name (single element character vector)}

\item{\code{tree_desc}}{numeric matrix with three columns: height, canopy radius,
  and measure of resource use (e.g. canopy surface area)}

\item{\code{max_height}}{maximum attainable height}

\item{\code{height_yr1}}{initial height for trees in a seedling cohort}

\item{\code{height_dbh_pars}}{numeric vector of length 3 giving the parameter values
  for the function relating trunk diameter (dbh) to tree height:
  \preformatted{dbh = exp(b1 + (b2 / (b3 + height)) )}}

\item{\code{height_fire_pars}}{numeric vector of length 4 giving the parameters for
  the generalized logistic function relating the proportional reduction in tree
  height for survivors of a fire:
  \preformatted{prop.reduced = 1 / (1 + exp(-(b1 + b2 * fi + b3 * h)))^b4}
  where \code{fi} is fire intensity and \code{h} is current height.}

\item{\code{growth_rate}}{rate of increase in height under ideal conditions (see Tree Growth).}

\item{\code{growth_rainfall_pars}}{numeric vector of length 3 giving the parameters
  for the logistic function relating rainfall and current tree height to
  proportional decrease in ideal growth rate (see Tree Growth).}

\item{\code{growth_crowding_pars}}{numeric vector of length 3 giving the parameters
  for the logistic function relating total stand resource use and current tree height to
  proportional decrease in ideal growth rate (see Tree Growth).}

\item{\code{coppice_boost_pars}}{numeric vector of length 3 giving the parameters for
  the function relating boosted growth rate of coppiced trees to height before
  coppicing (see Tree Growth).}

\item{\code{max_sprogs}}{single numeric value for maximum number of trees in a new cohort
  (ie. when space is unrestricted)}

\item{\code{seed_output}}{numeric matrix with three columns: height (e.g. in whole metres);
  seeds per tree; and retained proportion of seed.}

\item{\code{seed_rainfall.pars}}{numeric vector of length 2 giving the parameters for
  the function relating proportional adjustment of seed production to rainfall
  (see Seed)}

\item{\code{recruit_canopy_fn}}{an optional function to transform stand canopy area
  into the area perceived by recruits of this species (ie. to increase or reduce
  the inhiting effect of overstorey on recruitment). The function should take
  one argument for canopy area and return a single numeric value for transformed
  area.}

\item{\code{external_seed_fn}}{an optional function to return the number of seeds
of this species entering the stand from external sources in a given year.
The function should take no arguments and return a single numeric value.}
}}
\section{Tree growth}{

Changes in tree dimensions are modelled in a very basic fashion. change in tree height
is modelled as a function of existing height, moisture (rainfall), and stand resource
use. All other dimensions such as trunk diameter, canopy radius and canopy surface area,
are derived deterministically from tree height.

Annual growth in tree height is described by:
\preformatted{dh = (hmax - height) * g.adj}
where
\code{dh} is the increment in height;
\code{hmax} is the maximum attainable height for the species;
\code{height} is current height; and \code{g.adj} is adjusted growth rate:
\preformatted{g.adj = g * g.rain.adj * g.crowd.adj}
where
\code{g} is the ideal growth rate;
\code{g.rain.adj} is the proportional adjustment due to rainfall effect;
\code{g.crowd.adj} is the proportional adjustment due to crowding (resource
competition).

The rainfall and crowding adjustments to growth are given by generalized logistic
functions:
\preformatted{g.rain.adj = 1 / (1 + exp( -(b1 + b2 * rain * height^b3)))}
\preformatted{g.crowd.adj = 1 / (1 + exp( -(c1 + c2 * combined.resource.use * height^c3)))}

\code{combined.resource.use} is the summed resource use for all extant cohorts.

Trees which have been coppiced have a boosted growth rate which is meant to represent
their more extensive root system compared to un-coppiced trees of the same height.
\preformatted{g.boosted = max( g, b1 / (1 + b2 * exp(-b3 * hprev)) )}
}

\section{Seed}{

Seed production of trees is modelled as a function of height, based on a lookup
matrix with columns: tree height, seed production, proportion retained.
}

