% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tm.site.R
\name{tm.site}
\alias{tm.site}
\title{Simulates dynamics of a woodland stand.}
\usage{
tm.site(Spp, initial.cohorts, rain, scheduled.fires = NULL,
  fire.func = NULL, fire.canopy.func = NULL, fire.early.prob = 0,
  overlap.matrix = 0, ov.gen = 0, recruitment = TRUE, thinning = NULL,
  special = NULL, seedling.survival = NULL, session.settings = NULL,
  database = NULL)
}
\arguments{
\item{Spp}{either a single object or a \code{list} of objects
  of class \code{\linkS4class{SpeciesParams}} defining one or more species}
}
\description{
\code{tm.site} simulates the dynamics of cohorts of one or more tree species
in a woodland stand. For each cohort the simulation tracks annual values of
age, number of trees, height, plus a set of morphometrics derived from height
such as DBH and canopy radius. The location of individual trees is not
explicitly modelled. The stand environment is defined by input vectors for
annual rainfall and fire occurrence / intensity. The simulation period is
equal to the length of the rainfall vector.
}
\details{
The order of events in each year of the simulation is:
\enumerate{
  \item Optional special actions (coppice ON)
  \item Optional thinning of trees
  \item Growth of trees (uniform within cohort)
  \item Survival (adjustment of numbers within cohorts)
  \item Recruitment (addition of new cohorts)
  \item Fire (possible mortality and reduction of tree height)
  \item Report cohort statistics
}
}

