% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tmdbGetBlob.R
\name{tmdbGetBlob}
\alias{tmdbGetBlob}
\title{TODO - remember what this one does !}
\usage{
tmdbGetBlob(tmdb, paramSetID = NULL)
}
\arguments{
\item{tmdb}{an open database connection}

\item{paramSetID}{identifier of the parameter set to retrieve,
  or NULL (default) to be prompted}
}
\description{
TODO - remember what this one does !
}

