% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tmdbGetParamSet.R
\name{tmdbGetParamSet}
\alias{tmdbGetParamSet}
\title{Gets simulation params from a \code{tmrun} output database}
\usage{
tmdbGetParamSet(tmdb, paramSetID = NULL)
}
\arguments{
\item{tmdb}{an open database connection}

\item{paramSetID}{identifier of the parameter set to retrieve,
  or NULL (default) to be prompted}
}
\value{
a named \code{list} of runtime parameters
}
\description{
Gets simulation params from a \code{tmrun} output database
}

