% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tmdb_utility_functions.R
\name{tmdbOpen}
\alias{tmdbOpen}
\title{Open an output database from the \code{\link{tmRun}} function.}
\usage{
tmdbOpen(path = NULL, driver = NULL)
}
\arguments{
\item{path}{the database file path (if NULL a
  dialog will be displayed to choose the file)}

\item{driver}{an optional database driver to manage
  the connection}
}
\value{
the database connection
}
\description{
Attempts to open a connection to the given database and, if
successful, checks that the database contains the expected tables
by calling \code{\link{tmdbValidate}}.
}
\examples{
\dontrun{
con <- tmdbOpen("somedir/results.db")

# do interesting things, then...

tmdbClose(con)
}
}
\seealso{
\code{\link{tmdbOpen}}, \code{\link{tmdbValidate}}
}

