% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tmdbPlot.R
\name{tmdbPlot}
\alias{tmdbPlot}
\title{Plots yearly summed value (across cohorts by species) for a chosen variable.}
\usage{
tmdbPlot(tmdb, val = c("mergedarea", "basalarea", "corearea", "resourceuse",
  "n"), runid = 1, col = c("blue", "green", "purple", "magenta", "cyan"),
  x.same = TRUE, y.same = TRUE, show.legend = TRUE,
  pos.legend = c("left", "right"), plot.total = TRUE, col.total = "BLACK")
}
\arguments{
\item{tmdb}{an open connection to a \code{\link{tmRun}} output database}

\item{val}{name (character) of the variable to plot. The values plotted
  will be the sum of individual cohort values for each year.}

\item{runid}{a vector of one or more run ID values, each of which will
  be used to retrieve data for a separate plot}

\item{col}{vector of colours for species lines}

\item{x.same}{if TRUE, all plots use the same xlim setting}

\item{y.same}{if TRUE, all plots use the same ylim setting}

\item{show.legend}{if TRUE, legend is added to each plot}

\item{pos.legend}{specifies whether to plot the legend at upper-left or upper-right
  of plot area (default is left)}

\item{plot.total}{if TRUE, and there are two or more species in the results,
  plot the combined value in addition to that for each species}

\item{col.total}{colour to use for plotting total value}
}
\description{
Plots yearly summed value (across cohorts by species) for a chosen variable.
}

