% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tmdbSummary.R
\name{tmdbSummary}
\alias{tmdbSummary}
\title{Generate a summary of the contents of a \code{\link{tmRun}} output database.}
\usage{
tmdbSummary(tmdb)
}
\arguments{
\item{tmdb}{an open database connection}
}
\value{
a `data.frame` of summary data describing the simulation parameters
}
\description{
Generate a summary of the contents of a \code{\link{tmRun}} output database.
}
\examples{
\dontrun{
con <- tmdbOpen("results.db")
dat <- tmdbSummary(con)
tmdbClose(con)
}
}

