% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tmdb_utility_functions.R
\name{tmdbValidate}
\alias{tmdbValidate}
\title{Check that a database of simulation outputs is valid.}
\usage{
tmdbValidate(tmdb, show.output = TRUE)
}
\arguments{
\item{tmdb}{an open database connection}

\item{show.output}{TRUE (default) to print summary of database contents}
}
\value{
TRUE if the database was open and contained the required tables;
  FALSE otherwise
}
\description{
Checks that a database contains the expected tables for \link{\code{tmRun}}
output and, optionally, prints a summary of the database contents.
}
\seealso{
\code{\link{tmdbOpen}}
}

