% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/woodland.R
\docType{package}
\name{woodland}
\alias{woodland}
\alias{woodland-package}
\title{woodland: A package for woodland stand simulation.}
\description{
Blah blah blah
}
\section{Simulation function}{


Some words about running a simulation with \link{\code{tmRun}}...
}

\section{Utility functions}{


Some words about tmdbOpen and friends used with output SQLite databases.
}

